Move class to api
